// src/app/models/user-registration.model.ts
export class UserRegistration {
    companyName: string = '';
    industryId: number = 0;
    userName: string = '';
    firstName: string = '';
    login: string = '';
    password: string = '';
    passwordConfirmation: string = '';
    email: string = '';
    termsAccepted: boolean = false;
    privacyPolicyAccepted: boolean = false;
  }
  